Address PR feedback: project use subcommand, drop analytics annotation
- Restructure to 'tiger project use <project-id>' (alias: switch), leaving
  'tiger project' as a plain group command so future subcommands fit the
  namespace convention
- Set SilenceUsage: true on the command struct, matching the pattern from #194
- Remove the OmitArgsAnnotation machinery: every analytics event already
  carries the active project_id, so scrubbing the positional arg added no
  protection
- Inline the project-list fetch at both call sites, keeping only
  requireProjectAccess shared (it owns the stderr-only ID echo and exit code)
