+++ internal/cmd/root.go
@@ -131,15 +131,11 @@ func wrapCommands(cmd *cobra.Command, app *common.App, skipUpdateCheck *bool) {
 			// credentials it just stored.
 			start := time.Now()
 			defer func() {
-				trackedArgs := args
-				if c.Annotations[analytics.OmitArgsAnnotation] == "true" {
-					trackedArgs = nil
-				}
 				cfg, client, projectID := app.TryGetAll()
 				a := analytics.New(cfg, client, projectID)
 				a.Track(
 					fmt.Sprintf("Run %s", c.CommandPath()),
-					analytics.Property("args", trackedArgs),
+					analytics.Property("args", args), // NOTE: Safe right now, but might need allow-list in the future if some args end up containing sensitive info
 					analytics.Property("elapsed_seconds", time.Since(start).Seconds()),
 					analytics.FlagSet(c.Flags()),
 					analytics.Error(runErr),
